moved comment, thread and profile folder into reusable
