+++ src/__tests__/Components/reusable-components/CommentCardComponents/CommentCard.test.js
@@ -1,8 +1,8 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { CommentCard } from '../../../Components/reusable-components/CommentCardComponents/CommentCard';
+import { CommentCard } from '../../../../Components/reusable-components/CommentCardComponents/CommentCard';
 import '@testing-library/jest-dom/extend-expect';
-import { arrayOfUsersWhoLiked } from '../../../__mocks__/index';
+import { arrayOfUsersWhoLiked } from '../../../../__mocks__/index';
 
 describe('<CommentCard />', () => {
   it('should render', () => {
