+++ src/__tests__/Components/reusable-components/CommentCardComponents/NewCommentCard.test.js
@@ -1,8 +1,8 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { NewCommentCard } from '../../../Components/reusable-components/CommentCardComponents/NewCommentCard';
+import { NewCommentCard } from '../../../../Components/reusable-components/CommentCardComponents/NewCommentCard';
 import '@testing-library/jest-dom/extend-expect';
-import { user } from '../../../__mocks__/index';
+import { user } from '../../../../__mocks__/index';
 
 describe('<NewCommentCard />', () => {
   it('should render', () => {
