+++ src/__tests__/Components/reusable-components/CommentCardComponents/UpdateComment.test.js
@@ -1,8 +1,8 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { UpdateComment } from '../../../Components/reusable-components/CommentCardComponents/UpdateComment';
+import { UpdateComment } from '../../../../Components/reusable-components/CommentCardComponents/UpdateComment';
 import '@testing-library/jest-dom/extend-expect';
-import { text } from '../../../__mocks__/index';
+import { text } from '../../../../__mocks__/index';
 
 describe('<UpdateComment />', () => {
   it('should render', () => {
