+++ src/__tests__/Components/reusable-components/ProfileCardBComponents/ProfileCardOrgsField.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { ProfileCardOrgsField } from '../../../Components/reusable-components/ProfileCardComponents/ProfileCardOrgsField';
-import { orgs, user } from '../../../__mocks__/index';
+import { ProfileCardOrgsField } from '../../../../Components/reusable-components/ProfileCardComponents/ProfileCardOrgsField';
+import { orgs, user } from '../../../../__mocks__/index';
 import '@testing-library/jest-dom/extend-expect';
 
 describe('<ProfileCardOrgsField />', () => {
