+++ src/__tests__/Components/reusable-components/ProfileCardBComponents/ProfileCardUserRow.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { ProfileCardUserRow } from '../../../Components/reusable-components/ProfileCardComponents/ProfileCardUserRow';
-import { user } from '../../../__mocks__/index';
+import { ProfileCardUserRow } from '../../../../Components/reusable-components/ProfileCardComponents/ProfileCardUserRow';
+import { user } from '../../../../__mocks__/index';
 import '@testing-library/jest-dom/extend-expect';
 
 describe('<ProfileCardUserRow />', () => {
