+++ src/__tests__/Components/reusable-components/ProfileCardBComponents/ProfleCardButton.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { ProfileCardButton } from '../../../Components/reusable-components/ProfileCardComponents/ProfileCardButton';
-import { user } from '../../../__mocks__/index';
+import { ProfileCardButton } from '../../../../Components/reusable-components/ProfileCardComponents/ProfileCardButton';
+import { user } from '../../../../__mocks__/index';
 import '@testing-library/jest-dom/extend-expect';
 
 describe('<ProfileCard />', () => {
