+++ src/__tests__/Components/reusable-components/ThreadCardComponents/ThreadLeftComponentText.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import ThreadLeftComponentText from '../../../Components/reusable-components/ThreadCardComponents/ThreadLeftComponentText';
-import { heading, info } from '../../../__mocks__/index';
+import ThreadLeftComponentText from '../../../../Components/reusable-components/ThreadCardComponents/ThreadLeftComponentText';
+import { heading, info } from '../../../../__mocks__/index';
 import '@testing-library/jest-dom/extend-expect';
 
 describe('<ThreadLeftComponentText />', () => {
