+++ src/__tests__/Components/reusable-components/ThreadCardComponents/ThreadMiddleComponent.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import ThreadMiddleComponent from '../../../Components/reusable-components/ThreadCardComponents/ThreadMiddleComponent';
-import { heading, info } from '../../../__mocks__/index';
+import ThreadMiddleComponent from '../../../../Components/reusable-components/ThreadCardComponents/ThreadMiddleComponent';
+import { heading, info } from '../../../../__mocks__/index';
 import '@testing-library/jest-dom/extend-expect';
 
 describe('<ThreadMiddleComponent />', () => {
