+++ src/__tests__/Components/reusable-components/ThreadCardComponents/ThreadRightComponent.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { ThreadRightComponent } from '../../../Components/reusable-components/ThreadCardComponents/ThreadRightComponent';
-import { orgs, user } from '../../../__mocks__/index';
+import { ThreadRightComponent } from '../../../../Components/reusable-components/ThreadCardComponents/ThreadRightComponent';
+import { orgs, user } from '../../../../__mocks__/index';
 import '@testing-library/jest-dom/extend-expect';
 
 describe('<ThreadRightComponent />', () => {
